Move some related utils to fonts package
